Update ui-carts
- Change type of id to long
- Update entity model
- Update ui-carts module

+++ ui/ui-detail/src/main/java/com/study/compose/ui/detail/interactor/intent/DetailIntent.kt
@@ -3,7 +3,7 @@ package com.study.compose.ui.detail.interactor.intent
 import com.study.compose.ui.detail.data.ProductDetail
 
 sealed class DetailIntent {
-    data class Initial(val productId: Int) : DetailIntent()
+    data class Initial(val productId: Long) : DetailIntent()
     object GetProducts : DetailIntent()
     data class AddCart(val product: ProductDetail, val amount: Int) : DetailIntent()
 }
